requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
build==1.2.1
    # via pip-tools
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
idna==3.7
    # via requests
packaging==24.1
    # via build
pip-tools==7.4.1
    # via -r requirements.in
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.1
    # via -r requirements.in
requests==2.32.2
    # via -r requirements.in
semver==3.0.2
    # via -r requirements.in
urllib3==2.2.2
    # via requests
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
